refactor: Split out set tests.
